Revert "Merge pull request #7 from Joshua-ona/Essie/desktop"
This reverts commit dd1ada8c07bd8ebdccfb37a7db76675bb560c2ee, reversing
changes made to df444a6c0b62b6661394316817b5c53516931cd5.

+++ desktop/E-Discuss/src/main/java/com/edudiscuss/models/ActiveQuizResponse.java
@@ -5,10 +5,10 @@ public class ActiveQuizResponse {
     private boolean active;
     private int quiz_id;
     private String deadline;
-    private double remaining_seconds;
+    private int remaining_seconds;
 
     public boolean isActive() { return active; }
     public int getQuizId() { return quiz_id; }
     public String getDeadline() { return deadline; }
-    public double getRemainingSeconds() { return remaining_seconds; }
+    public int getRemainingSeconds() { return remaining_seconds; }
 }
